Rework tests, add scripts
