+++ package.json
@@ -8,13 +8,15 @@
     "start": "",
     "sandbox": "ng serve",
     "build": "gulp inline && ngc -p \"./tsconfig.module.json\" && gulp cleanup",
-    "test": "ng test",
-    "e2e": "ng e2e"
+    "test": "ng test --watch=false --code-coverage",
+    "e2e": "ng e2e",
+    "preversion": "ng lint",
+    "postpublish": "git push"
   },
-  "main": "dist/inline/index.js",
-  "types": "dist/inline/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
-    "dist/inline"
+    "dist"
   ],
   "peerDependencies": {
     "@angular/animations": "^4.0.0 || ^5.0.0",
@@ -40,6 +42,7 @@
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
+    "@types/prismjs": "^1.9.0",
     "angular2-notifications": "^0.9.7",
     "codelyzer": "^4.0.1",
     "core-js": "^2.4.1",
