--- src/app/mock/mock.component.ts
@@ -1,7 +1,0 @@
-import {Component} from "@angular/core";
-
-@Component({
-  selector: 'mock-component',
-  template: '<code codeHighlight>WELCOME TO CODE</code>'
-})
-export class MockComponent {}
